requirements.txt
aiocache >= 0.11.1
aiofiles >= 0.4.0
aiohttp >= 3.5.4
aiomysql >= 0.0.20
beautifulsoup4 >= 4.8.0
bs4 >= 0.0.1
discord.py >= 1.2.5
emoji >= 0.5.3
html5lib == 1.0.1
pydle >= 0.9.3
pyz3r >= 4.2.0
pysmz3 >= 1.0.1
pyyaml >= 5.1.2
ircmessage
Quart
quart-openapi
gspread_asyncio
oauth2client
jishaku
python-dotenv
html2markdown